Hardcore add_contract paths hack to force compiling with rebar3
rebar3 doesn't switch into each dependency directory when compiling
so when ubf is added as a dependency for other apps using rebar3,
you'll get compile errors as the contracts will not be found. This
is a hack to hardcode the usual ubf path into the files so that it
will compile.